Update create node pool arguments in cluster service
Currently the function expects a cluster options model and a cluster
name as a string. The cluster options is unnecessary because that model
describes how to create the cluster, not the node pool.

This changes it so that the function expects a cluster model which
contains information about the cluster that has been created.