Replace bind with bind json in gin handlers
The current behaviour of using the bind method means that gin checks the
content type header and parses the request body dynamically. But
inconsistently there are already some handlers which use the bind json
method.

This uses the bind json method consistently to ensure that only json can
be used in all handlers.