Add delete garbage method to rancher package
Used to delete rancher resources that cannot be deleted until they are
no longer in use. For example, node templates.

Uses a mutex and map to store garbage objects. This is to ensure that
multiple goroutines does not try to add objects at the same time.

Because the garbage deletion method locks the mutex and blocks all
functions trying to add garbage it needs to run as fast as possible. No
retries are attempted during a delete run.