Add encoding and decoding scheme for organizations in rancher
To allow multiple organizations to create clusters with the same name in
rancher.

Because of the limited characters of rancher cluster name a dash is used
to separate organization from cluster name. Since users might have
dashes in both organization and cluster names these are double encoded
to avoid collisions.

The unit test for the new functions can be read as reference as how the
encoding and decoding behaves.