Update create node pool to get closest flavor in rancher
Current behaviour requires the cpu, disk and cpu count to exactly match
a flavor available in openstack. When a flavor cannot be matched
exactly the user will get a internal server error response.

This uses the closest flavor method to get the most suitable one based
on the user input.