Update new rancher function to use options pattern in rancher
Current behaviour for specifying options for rancher makes the type
signature of the new rancher function longer the more options are added.

This changes it so that the functions options pattern is used instead.
This allows for future options to easily be added in the future.

This is not a standard pattern described in upstream documentation but
rather a community convention.

How this pattern work might not be obvious at a first glance but there
are several resources such as:

- https://github.com/tmrts/go-patterns/blob/master/idiom/functional-options.md
- https://golang.cafe/blog/golang-functional-options-pattern.html