Add helm related fields to app model
Adds the necessary fields to be able to specify repository and release
for a helm chart.

The helm values has to be treated a bit special. In order to support
values that will be unique per chart a map of strings to any is used.

In this initial version the helm values are just stored, even though the
input might be invalid.