Update app model to store helm values as json in database
Because the helm values is a recursive map the database driver can not
store it in a column and makes no attempt to convert it resulting in an
error.

This adds the standard scan and value methods to the helm values type to
convert the map into something that the database driver supports.

The value method takes a helm values type and uses json marshalling to
convert it into a list of bytes that the database driver can store.

The scan method does the reverse and uses json unmarshalling to take the
list of bytes from the database and convert it into a helm values type.