Add bootstrapping helpers to rancher package
These helpers use the bootstrap and internal ca secrets created by
rancher. The limitation is that dockyards backend must be running in the
same kubernetes cluster as rancher.

The helpers creates a token and sets a new password for the admin user.

Both the token and the new password is saved in a rancher credentials
secret in the dockyards namespace.