Update get kubeconfig to generate limited tokens in rancher
Currently when generating a kubeconfig for a rancher cluster the
resulting token is unrestricted and never expires. For a malicous user
this token can be used to access any cluster in rancher.

Unfortunately it is not possible to request a kubeconfig from the
rancher api specifying limitations on the resulting token.

This updates get kubeconfig to:
- Generate a kubeconfig for a specific cluster
- Generate a token scoped to a specific cluster and with a short ttl
- Update the kubeconfig to use the new limited token
- Delete the unrestricted token created automatically by rancher