Update get kubeconfig to return config in cluster service interface
Currently the get kubeconfig method returns a string of a marshalled
config object. If other parts needs to access parts of the kubeconfig it
first needs to be unmarshalled and then remarshalled before returning it
to the user.

This updates the method to return the raw config object and any user
facing functions only needs to marshal the config once.