Update require auth middleware to store subject in gin context
Currently the require auth middleware lookup up and stores a full user
object from the database in the gin context.

This updates the middleware to only store the sub claim from the
supplied token. The users are no longer stored in the database and it
has always been a very unnecessary action to fetch the user on every
request.