Update main to create separate manager and controller client
Currently the client from the manager is passed to handlers and metrics.
This client can not be used before the manager has been started and
synced caches causing a race condition for components that needs to
fetch configuration items from kubernetes objects.

This created a separate controller client to be used during
configuration.