Update organization to be cluster scoped in v1alpha1
Currently the organization is namespace scoped. This implies that
multiple instances of dockyards could be running in the same cluster
with different contexts. This also requires all controllers to be
context aware.

This changes the organization type to be cluster scoped so that
organization names are unique to the entire cluster.