Update get cluster kubeconfig handler to generate kubeconfig
The current method of serving the cluster-api generated kubeconfig to
the user has server drawbacks. The kubeconfig contains the canonical
endpoint, which is not always the endpoint that the user needs to
connect to. The kubeconfig also has a validity period of 365 days.

This updates the handler to generate a kubeconfig using the cluster
endpoint and a short validity.

DY-214 #done.