Update api handler to embed kubernetes client field
Currently the client field is name controllerClient as a remnant from
when the api handler had multiple clients to databases and cloud
environments. After kubernetes is the only datasource that the api
handlers manage this make the code unnecessarily complex.

This update the kubernetes client to be an embedded field to simplify
the code readability.