Update create cluster handler to use cluster templates directly
Currently the create cluster handler uses the get recommended node pools
helper to convert a cluster template to node pool options. This requires
that every new field added to the node pool type is also supported in
the helper.

This replaces the use of the helper with a deep copy of the node pools
in the cluster template.