Fix inconsisent path value in create cluster handler
Currently the create cluster handler expects an organization name but
there are several instances of the variables being called organization
id. This is a leftover from when the organization id was the expected
value.

This corrects the inconsistencies and uses variables called organization
name everywhere. For further consistency the method itself is renamed to
create cluster.