Update v1alpha1 types to be be marked as unserved versions
Previously the v1alpha1 version of all types were removed too early.
After a previous commit reverted the deletion this adds a marker to all
v1alpha1 types to stop serving them. They will exist in kubernetes but
generate errors if requested. This is to allow all resources to be
migrated to use the latest storage version. Only when all resources have
been updated to use the latest storage version can the v1alpha1 versions
of types be removed completely.