Update get kubeconfig handler to require cluster patching permission
Currently it is enough to have get permissions on a cluster resource to
download a kubeconfig. This results in users with only reader permission
being able to download a kubeconfig giving the full permission to the
kubernetes api.

This updates the get kubeconfig handler to require patch permissions on
cluster resources to download the kubeconfig. This blocks users with
only reader permission to download kubeconfigs.

A long term solution will be to implement support for generating a
read-only kubeconfig.