Add util for creating a test environment
In order to properly test subject access reviews the fake client can not
be used. The controller-runtime developers recommends that the envtest
package is used but this is more complicated to configure.

This adds a helper for creating a test environment using envtest. The
helper function configures envtest but also creates the basic dockyards
resources needed to test most functionality.