Update validating webhooks to use reader client
Currently the validating webhooks uses the feature package which tracks
all enabled features using a separate controller. This makes it hard to
use and test the webhooks in isolation.

This updates the validating webhooks to use a reader client to directly
interact with kubernetes. This follows the pattern established by
upstream projects like cluster-api.