Add reference name field to status in worktree v1alpha3 type
Currently the worktree type has a reference to the url of a repository
that has reconciled the file contents. This causes issues with tools
that requires a reference to clone. It also creates an implicit
relationship between components which does not necessarily agree on what
the default branch of a git repository should be.

This adds an optional reference name field to the wortree status to
allow components to explicitly set or expect a repository reference.

Relates to DY-390.