Update create workload handler to return details when unprocessable
Currently when the create workload handler gets an invalid error from
kubernetes only the unprocessable entity status code is returned. This
makes it very complicated for users to troubleshoot what was invalid
about the request.

This updates the create workload handler to extract the status error if
possible and return the detailed causes for the rejection.