Update api handler tests to use retry until found helper
Currently the api handler tests have a hardcoded sleep to wait between
test resources are created and the handlers are called. These are needed
to wait for the client cache to be populated. The hardcoded sleep is too
long with fast compute resources but not long enough with slow compute
resources, i.e cloud builds.

This updates the api handler tests to use the retry until found helper
to improve the wait times.