Update create cluster workload handler to support optional namespace
Currently the create cluster workload handler requires that the workload
target namespace is set. Often this will be set to the same value as the
name. The behaviour of the create cluster deployment was to use the name
as target namespace when missing.

This updates the create cluster workload handler to use the name of the
workload as the target namespace when not explicitly set.

Relates to DY-398.