Add references field to workload v1alpha3 type
Intended to be used to store references to all objects reconciled in the
target clusters.

Each entry stores a reference to a parent, i.e the object in the
management cluster that reconciles the entry.

Each entry stores a list of urls for that object to simplify summarizing
all urls of the workload.

This supersedes the dependency references field.

Relates to DY-396.