Fix incorrect use of pointer for embedded field in workload reference
Currently the embedded field typed object reference is a pointer. It
seems that the controller tools does not take this into account and has
generated the custom resource definition to require the kind and name
fields. But it being a pointer indicates that is would be valid to omit
it.

This updates the embedded field to not be a pointer to make it more
semantically clear that it is not an optional field.