Add generic helper for create cluster resource handlers
Currently all handler methods have a lot of duplicated code for
authorization and encoding including error handling.

This adds a generic helper for create cluster resource handlers. The
majority of the duplicated code is made redundant as the generic helper
handles authorization and the encoding and decoding of the request and
the response.

The updated create cluster resources handlers only has to implement the
actual logic for validating and creating the kubernetes resources.