Update the generic update helper to return accepted status code
Currently the generic update helper uses the no content status code on
success. This does not really indicate well that the update has only
been accepted and that nothing has necessarily been finalized.

This updates the status code to be accepted to better indicate the lack
of any guaranteed results.