Add identity provider types and api endpoint
Representing identity providers is the first step in supporting single
sign-on. This version of the IdentityProvider object is intended to have
the fields sufficient for OIDC authentication with any compliant
provider.

The /v1/identity-providers should supply enough information for a
front-end to display available SSO login possibilities.

Many OIDC providers supply a publicly available endpoint to assist in
configuration. See:
https://openid.net/specs/openid-connect-discovery-1_0.html

A dockyards IdentityProvider is expected to have either an
oidcProviderDiscoveryURL or a full oidcProviderConfig. An object can be
created with neither, but will be omitted from the output of
/v1/identity-providers.

Bumped the dockyards-api version to make use of the new type defined
there.