Update list global organization handler tests to sort slices
The lists returned by the global organization handler does not do any
sorting by design. This is fine for the intended use case but makes the
unit tests randomly fail when the exact order does not match the
expected result.

This updates the list global organization handler tests to sort the
result during comparison.