Add namespace reference to v1alpha3 organization spec
Currently the namespace reference of an organization is stored in the
organization status. This causes issues during creation since status can
only be set on existing objects.

This adds a namespace reference field to the organization spec to allow
an organization to be created with an existing namespace reference. This
deprecates the namespace reference field in the organization status but
keeps it for backwards compatibility.