Add provider id field to v1alpha3 node type
Currently the underlying id of a node is stored in the cloud service id
status field. The word cloud being used is poorly chosen implying that
the field would only be set for nodes in cloud environments.
Additionally, the field being part of the status subresource makes it
impossible to set during object creation.

This adds a provider id field to the node spec. This supersedes the
cloud provider id status field.