Update validate json middleware to embed cue sources
Currently loading the cue sources for the validate json middleware is
complicated to use due to the requirement of passing in the working
directory.

This update the validate json middleware to embed the cue sources
allowing it to be loaded as an overlay removing the need for consumers
to specify the working directory.