Add shared suite for handler tests
Currently all handler tests are required to prepare their own testing
environments. This causes a lot of duplicated code but also longer
running tests as every test has to wait for the environment to start.

Also, most handler tests currently invoke the handler functions directly
instead of using the registered routes. This makes the tests not really
test or catch issues with the configured patterns.

This adds a shared test suite to setup and teardown a test environment
once that all handler test then use. The shared suite also creates a
request multiplexer with the real routes registered.