Move api package to root of project to follow upstream conventions
This moves the api package to the root of the project to follow upstream
conventions. Kubernetes and kubebuilder specifically has established the
conventions that all custom resources belongs in the api package at the
root of the project.

As part of moving the repository hosting to github other components must
update their import paths and this is the opportunity to fix these kinds
of inconsistencies.