Update cluster controller upgrade tests to wait for expected state
Currently the cluster controller upgrade tests only wait for the cluster
upgrades ready condition. This causes random failures in slower
execution environments since the cluster reconciler might not be done by
the time the tests occur.

This updates the cluster controller upgrade test to poll for the actual
expected state rather than just the condition to ensure that the
reconciliation is actually done.