Update create global token handler tests to wait for condition
Currently the create global token handler tests only waits for the user
resources to exist. This causes issues on slow execution environments
where the user will exist but not with the expected conditions.

This updates the tests to wait for the conditions to be available before
executing any of the test cases.