Update cluster controller to support cluster upgrades feature
Currently the cluster controller will always reconcile cluster upgrades
even when the feature is not enabled. This makes it misleading when
cluster upgrades are listed as available but not allowed.

This updates the cluster controller to always reconcile cluster upgrades
to an empty list when the feature is not enabled.