Update webhooks to use default service
Currently the webhooks uses the old controller-gen behaviour of having
static validating webhook configurations and services names. These
require a complicated set of kustomization patches to get unique names.

This updates the webhooks to set the service name markers supported by
the latest version of the controller-gen tool. This simplifies the
configuration manifests and kustomization patches.

This removes the dedicated webhook service as it is made redundant.