Update /refresh to set Content-type header to JSON
Currently `/refresh` returns a body formatted as JSON but sets the
`Content-Type` header to `text/plain`. This does not allow with the
OpenAPI spec in sudoswedenab/dockyards-api, which means that the
generated client cannot be used correctly in sudoswedenab/dockyardsctl.

This commit ensures that `Content-Type` is set to `application/json` as
per OpenAPI spec