Update create test environment helper to not set display name
Currently the organizations created by the test environment helper will
have a display name set by default. This results in a lot of tests
unrelated to organization information having to ensure correct display
names.

This changes the behaviour to not set a display name by default. Only
tests explicitly related to organization displays names needs to test
for it.