Add dockyards-external-url flag
Since the verificationrequest creation is moving to backend, we need to
be able to specify the verification URL in the backend. This means that
the backend needs to be aware of its own external name. Although there
exists a `dockyards-public` service that holds the ExternalName for the
installation, we might move away from it in future, so we will use a
flag instead.