Add User reconciler for signups
Currently the verificationrequest type is created by the backend, but
managed entirely by serverside; VRs are created on user creation.

We are moving certain areas of responsibility away from the serverside
to backend and VR reconciliation is one of them. This commit adds a
reconciler that watches for User changes and CRUDs VerificationRequests
accordingly