Add api/config sub-module for config management
As the configuration for Dockyards services grows more complex, we are
considering moving away from flags to using a "config file-like"
approach using a Kubernetes ConfigMap resource or similar.

This commit introduces a sub-module `dockyards-backend/api/config` that
(for now) is used to configure ExternalURL for Dockyards. This is done
in a submodule with the idea that other components can use the same
approach for configuration.