Add experimental v2 api
Currently the v1 api uses custom types that needs to be converted to and
from the actual kubernetes resources. This causes a lot of duplicate
work where everything must be implemented twice, first as kubernetes
resources and then as v1 equivalents.

This adds an experimental v2 api that uses the kubernetes types
directly. This is considered experimental as it in still unclear if this
is a good approach or not.

In this initial version the v2 support is limited to listing and getting
namespaced resources.