Update User reconciler to accept dockyards config as input
Currently the app fetches Dockyards Config on startup, extracts
`externalURL` key and passes it on to the User reconciler. While this
works fine, we will most likely want to reload Dockyards Config
dynamically in future, hence we will need to propagate the updates to
whatever reconcilers use the Dockyards Config. To make that simpler, we
can pass the reference to the DY Config to the reconcilers and let them
extract the keys themselves.

This commit updates User reconciler to expect a reference to the
DockyardsConfig and extract the key itself