Update get global token handler to reject deleting users
Currently users can refresh their access tokens while the user object is
being deleted. This is usually a very small window but if something is
blocking a user from being deleted it could causes issues.

This updates the get global token handler to reject any requests for
deleted users with an unauthorized error.