Update organization member handlers to use member types
This updates the list organization members handler to use the member
type. To preserve compatibility with the old behaviour the id of the
user is used for each member rather than the id of the member resource.

This also merges the special handler for users deleting their own
memberships into the delete organization member handler.