Add a default aggregate ClusterRole for authenticated users
We are looking into adding support for viewing shared resources that are
not cluster-scoped and are not owned by DY Backend.

This commit:
- adds a new ClusterRole `dockyards:authenticated` that is an aggregate
role to be used for granting users access to auxiliary resources, e.g.
WorkloadForms
- adds a new (User)Group `dockyards:authenticated`, and ensures that all
users are considered part of it (similar to `system:authorized`)

Note that this commit does not include tests specific to this
functionality due to some limitations with `envtest` ephemeral clusters
and aggregate ClusterRoles. These tests will be added at a later stage