Update cluster webhook to mutate kubernetes versions
Currently there is nothing that ensures that the kubernetes version of a
cluster is set on creation when using the v2 api. This causes issues
when users create clusters without specifying the version and control
planes can not be reconciled.

This adds defaulting behaviour to the cluster webhook to either use the
default release or reject clusters if no usable default release can
be found.