Fix Organization creation when using v1 API
This commit addresses the bug described in #89 .

When creating an org, a Member is created automatically for the user
that triggers the org creation. The Member is created without some
mandatory labels, so the Member webhook denies the creation. This commit
updates the labels to make sure the creation succeeds