Fix Organization/Member creation order
As reported in #89, it is not possible to create an organization through
the v1 API. The underlying issue is that on org creation a member is
created automatically. Member needs an Organization to exist to make
sure that it can be attached to it, but at the moment Member is created
before the Organization. This commit re-orders the operations, so the
Member creation should succeed