Remove identity provider discovery related tests
These tests are testing a v1 endpoint that doesn't exist, so they are
not needed. Fetching the list of identity providers will in the future
be done via the v2 endpoints anyway. Will add tests for OIDC at a later
stage though, when we have a better idea of _how_ we should test the
integration in a robust manner.